+++ bfabric_app_runner/src/bfabric_app_runner/app_runner/resolve_app.py
@@ -2,50 +2,40 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-import yaml
-from pydantic import ValidationError
-
 from bfabric.experimental.workunit_definition import WorkunitDefinition
-from bfabric_app_runner.specs.app.app_spec import AppSpec
-from bfabric_app_runner.specs.app.app_version import AppVersion
+from bfabric_app_runner.specs.app.app_spec import AppSpec, BfabricAppSpec
 
 if TYPE_CHECKING:
     from pathlib import Path
     from bfabric import Bfabric
+    from bfabric_app_runner.specs.app.app_version import AppVersion
+
+
+def _get_application_version_str(workunit_definition: WorkunitDefinition) -> str:
+    """Returns the application version string from the workunit definition."""
+    # TODO this should be more generic in the future about the key for the app version (should be handled in AppSpec)
+    # TODO logic to define "latest" version (should also be handled in AppSpec)
+    if "application_version" not in workunit_definition.execution.raw_parameters:
+        msg = "The workunit definition does not specify an application version."
+        raise ValueError(msg)
+    return workunit_definition.execution.raw_parameters["application_version"]
 
 
 def _resolve_app(versions: AppSpec, workunit_definition: WorkunitDefinition) -> AppVersion:
     """Resolves the app version to use for the provided workunit definition."""
-    # TODO this should be more generic in the future about the key for the app version (should be handled in AppSpec)
-    # TODO logic to define "latest" version (should also be handled in AppSpec)
-    if "application_version" not in workunit_definition.execution.raw_parameters:
-        raise ValueError("The workunit definition does not contain an application version.")
-    app_version = workunit_definition.execution.raw_parameters["application_version"]
-    # TODO graceful handling of invalid versions
-    if app_version in versions and versions[app_version] is not None:
-        return versions[app_version]
-    else:
+    version_str = _get_application_version_str(workunit_definition=workunit_definition)
+    if version_str not in versions or versions[version_str] is None:
         msg = (
-            f"application_version '{app_version}' is not defined in the app spec,\n"
+            f"application_version '{version_str}' is not defined in the app spec,\n"
             f" available versions: {sorted(versions.available_versions)}"
         )
         raise ValueError(msg)
-    return versions[app_version]
-
-
-def _load_spec(spec_path: Path, app_id: int, app_name: str) -> AppVersion | AppSpec:
-    # TODO the reason this exists is I don't want to refactor and complicate the CLI right now, however,
-    #      it is not fully clear if this is perfectly sound in all cases.
-    data = yaml.safe_load(spec_path.read_text())
-    try:
-        return AppVersion.model_validate(data)
-    except ValidationError:
-        return AppSpec.load_yaml(spec_path, app_id=app_id, app_name=app_name)
+    return versions[version_str]
 
 
 def load_workunit_information(
     app_spec: Path, client: Bfabric, work_dir: Path, workunit_ref: int | Path
-) -> tuple[AppVersion, Path]:
+) -> tuple[AppVersion, BfabricAppSpec, Path]:
     """Loads the app version and workunit definition from the provided app spec and workunit reference.
 
     :param app_spec: Path to the app spec file.
@@ -53,33 +43,31 @@ def load_workunit_information(
     :param work_dir: Path to the work directory.
     :param workunit_ref: Reference to the workunit (ID or YAML file path).
     :return app_version: The app version to use.
-    :return workunit_ref: Path to the workunit definition file. Can be used to reference the workunit in further
-        steps to avoid unnecessary B-Fabric lookups. (If the workunit_ref was already a path, it will be returned as is,
-        otherwise the file will be created in the work directory.)
+    :return bfabric_app_spec: The B-Fabric app spec, which specifies the app runner version to use.
+    :return workunit_ref: The path to the workunit definition file. If the workunit_ref was already a path, it will be
+        returned as is, otherwise the file will be created in the work directory.
     """
     workunit_definition, workunit_ref = _resolve_workunit_definition(client, work_dir, workunit_ref)
-    app_version = _resolve_app_version(app_spec, workunit_definition)
-    return app_version, workunit_ref
-
-
-def _resolve_app_version(app_spec: Path, workunit_definition: WorkunitDefinition) -> AppVersion:
-    app_parsed = _load_spec(
-        app_spec, workunit_definition.registration.application_id, workunit_definition.registration.application_name
+    app_spec = AppSpec.load_yaml(
+        app_spec,
+        app_id=workunit_definition.registration.application_id,
+        app_name=workunit_definition.registration.application_name,
     )
-    if isinstance(app_parsed, AppVersion):
-        app_version = app_parsed
-    else:
-        app_version = _resolve_app(versions=app_parsed, workunit_definition=workunit_definition)
-    return app_version
+    app_version = _resolve_app(versions=app_spec, workunit_definition=workunit_definition)
+    bfabric_app_spec = app_spec.bfabric
+    return app_version, bfabric_app_spec, workunit_ref
 
 
 def _resolve_workunit_definition(
     client: Bfabric, work_dir: Path, workunit_ref: int | Path
 ) -> tuple[WorkunitDefinition, Path]:
-    """Given a workunit reference (ID or YAML file path), loads the workunit definition and returns it.
+    """Load the specified workunit definition.
 
-    You get both the workunit definition and the path to the workunit definition file, which will be created if it
-    did not exist before in `work_dir`.
+    If the workunit_definition.yml file exists in the work_dir, it will be used, otherwise, it will be created.
+    :param client: The B-Fabric client to use for resolving the workunit.
+    :param work_dir: The directory where the workunit definition file will be created or found.
+    :param workunit_ref: Reference to the workunit (ID or YAML file path).
+    :return: A tuple containing the WorkunitDefinition and the path to the workunit definition file.
     """
     workunit_definition_file = work_dir / "workunit_definition.yml"
     workunit_definition = WorkunitDefinition.from_ref(workunit_ref, client, cache_file=workunit_definition_file)