+++ bfabric_app_runner/src/bfabric_app_runner/cli/app.py
@@ -9,7 +9,7 @@ from bfabric.experimental.entity_lookup_cache import EntityLookupCache
 from bfabric.utils.cli_integration import use_client
 from bfabric_app_runner.app_runner.resolve_app import load_workunit_information
 from bfabric_app_runner.app_runner.runner import run_app, Runner
-from bfabric_app_runner.cli.cmd_prepare import _write_workunit_makefile
+from bfabric_app_runner.prepare.makefile_template import render_makefile
 
 
 @use_client
@@ -24,9 +24,13 @@ def cmd_app_run(
     client: Bfabric,
 ) -> None:
     """Runs all stages of an app."""
-    # TODO doc
-    app_version, workunit_ref = load_workunit_information(app_spec, client, work_dir, workunit_ref)
-    _write_workunit_makefile(path=work_dir / "Makefile", app_ref=app_spec)
+    # TODO docstring
+    app_version, bfabric_app_spec, workunit_ref = load_workunit_information(app_spec, client, work_dir, workunit_ref)
+
+    # Render the workunit Makefile template
+    render_makefile(
+        path=work_dir / "Makefile", bfabric_app_spec=bfabric_app_spec, rename_existing=True, use_external_runner=False
+    )
 
     # TODO(#107): usage of entity lookup cache was problematic -> beyond the full solution we could also consider
     #             to deactivate the cache for the output registration
@@ -59,17 +63,22 @@ def cmd_app_dispatch(
     :param workunit_ref: Reference to the workunit (ID or YAML file path).
     :param create_makefile: If True, a Makefile will be created in the app directory.
     """
-
     work_dir = work_dir.resolve()
     # TODO set workunit to processing? (i.e. add read-only option here)
 
-    app_version, workunit_ref = load_workunit_information(app_spec, client, work_dir, workunit_ref)
+    app_version, bfabric_app_spec, workunit_ref = load_workunit_information(app_spec, client, work_dir, workunit_ref)
     with EntityLookupCache.enable():
         runner = Runner(spec=app_version, client=client, ssh_user=None)
         runner.run_dispatch(workunit_ref=workunit_ref, work_dir=work_dir)
 
     if create_makefile:
-        _write_workunit_makefile(path=work_dir / "Makefile", app_ref=app_spec)
+        # Render the workunit Makefile template
+        render_makefile(
+            path=work_dir / "Makefile",
+            bfabric_app_spec=bfabric_app_spec,
+            rename_existing=True,
+            use_external_runner=False,
+        )
 
 
 def _write_file_chmod(path: Path, text: str, mode: int) -> None: